Fix a caffe2-gloo dependency problem
Summary:
The problem:
Building caffe2 fails because the installed directory contains "anaconda".

The cause:
Compiling Gloo will generate a new config.h file in the binary folder.
If we put the original config.h in front, the compiler will complain "Expected GLOO_USE_CUDA to be defined".

~~~Switch the positions of the include folders can solve the problem.~~~

Function caffe2_include_directories in cmake/Utils.cmake is a little bit hacky. If the directory contains "anaconda", it will append the new include directory after existing include path. Otherwise it will insert the directory before the path. So in the first case, the directories are inserted in order, and in the latter one, they are inserted reversely.

The solution:
See the commit.

pietern #1121
Closes https://github.com/caffe2/caffe2/pull/1258

Reviewed By: Yangqing

Differential Revision: D5907167

Pulled By: houseroad

fbshipit-source-id: 2cb3916e7e0313ebc3be3d1666bfa14bbf479607