Use SmallVector for TensorImpl sizes and strides. (#13649)
Summary:
This removes dynamic allocations for sizes/strides for tensors with <= 5
dims. This should cover the most common tensor use cases; we use a lot
of 4D tensors in images (N, C, H, W) and LSTMs use tensors with 3 or fewer dims.

Benchmarking results can be found here:
https://gist.github.com/zou3519/ce4182722ae7e2a228bc8b57ae60b0e9
The quick summary is that this PR:
- makes aten LSTM's forward pass ~1ms faster and improves JIT lstm perf
  as well
- Tensor as_strided is now 200ns faster for dimensions <= 5
- at::empty performance is 200ns slower for dimensions > 5. For dims <= 5,
  there is no noticeable perf change.
- Variable ops are 200-500ns faster because Variables never used their
  sizes/strides fields in the first place.
Pull Request resolved: https://github.com/pytorch/pytorch/pull/13649

Differential Revision: D12950409

Pulled By: zou3519

fbshipit-source-id: 0bd87ec9f712ddc0d533a347d781e3a91a954b90