Better stringification of dispatch keys in error messages (#21809)
Summary:
Pull Request resolved: https://github.com/pytorch/pytorch/pull/21809

Many error messages show dispatch keys, for example when the dispatcher didn't find a kernel to dispatch to.
Previously, this was a string like "CPU" or "CUDA" for known backends and just an arbitrary number for other backends.

Now, tensor type id registration also registers a name for the dispatch key and shows that in the error messages.

There is no API change, just the error messages are better now.

Differential Revision: D15835809

fbshipit-source-id: 4f0c9d0925c6708b02d79c653a2fae75b6623bb9